+++ json/src/main/java/com/basistech/rosette/dm/jackson/AnnotatedTextMixin.java
@@ -113,6 +113,9 @@ public abstract class AnnotatedTextMixin {
     public abstract ListAttribute<CategorizerResult> getSentimentResults();
 
     @JsonIgnore
+    public abstract ListAttribute<CategorizerResult> getTopicResults();
+
+    @JsonIgnore
     public abstract ListAttribute<Entity> getEntities();
 
     @JsonIgnore