+++ json/src/main/java/com/basistech/rosette/dm/jackson/array/AnnotatedTextArrayMixin.java
@@ -115,6 +115,9 @@ public abstract class AnnotatedTextArrayMixin {
     public abstract ListAttribute<CategorizerResult> getSentimentResults();
 
     @JsonIgnore
+    public abstract ListAttribute<CategorizerResult> getTopicResults();
+
+    @JsonIgnore
     public abstract ListAttribute<Entity> getEntities();
 
     @JsonIgnore