+++ json/src/test/java/com/basistech/rosette/dm/json/plain/JsonTest.java
@@ -84,6 +84,7 @@ public class JsonTest extends AdmAssert {
     private TranslatedTokens spanishTranslation;
     private CategorizerResult categoryResult;
     private CategorizerResult sentimentResult;
+    private CategorizerResult topicResult;
     private AnnotatedText referenceTextOldEntities;
     private AnnotatedText referenceText;
     private Entity entity;
@@ -282,6 +283,12 @@ public class JsonTest extends AdmAssert {
         crBuilder.add(sentimentResult);
         builder.sentimentResults(crBuilder.build());
 
+        crBuilder = new ListAttribute.Builder<>(CategorizerResult.class);
+        topicResult = new CategorizerResult.Builder("basketball", 0.5)
+                .confidence(0.3).build();
+        crBuilder.add(topicResult);
+        builder.topicResults(crBuilder.build());
+
         referenceTextOldEntities = builder.build();
     }
 
@@ -478,6 +485,12 @@ public class JsonTest extends AdmAssert {
         depListBuilder.add(new Dependency.Builder("V", -1, 0).build());
         builder.dependencies(depListBuilder.build());
 
+        crListBuilder = new ListAttribute.Builder<>(CategorizerResult.class);
+        topicResult = new CategorizerResult.Builder("basketball", 0.5)
+                .confidence(0.3).build();
+        crListBuilder.add(topicResult);
+        builder.topicResults(crListBuilder.build());
+
         referenceText = builder.build();
     }
 
@@ -624,6 +637,8 @@ public class JsonTest extends AdmAssert {
         assertEquals(categoryResult, read.getCategorizerResults().get(0));
 
         assertEquals(sentimentResult, read.getSentimentResults().get(0));
+
+        assertEquals(topicResult, read.getTopicResults().get(0));
     }
 
     @Test
@@ -693,6 +708,8 @@ public class JsonTest extends AdmAssert {
         assertEquals("V", read.getDependencies().get(0).getRelationship());
         assertEquals(-1, read.getDependencies().get(0).getGovernorTokenIndex());
         assertEquals(0, read.getDependencies().get(0).getDependencyTokenIndex());
+
+        assertEquals(topicResult, read.getTopicResults().get(0));
     }
 
     @Test