+++ json/src/main/java/com/basistech/rosette/dm/jackson/RelationshipMentionMixin.java
@@ -40,6 +40,7 @@ public abstract class RelationshipMentionMixin {
                              @JsonProperty("source") String source,
                              @JsonProperty("confidence") Double confidence,
                              @JsonProperty("modality") String modality,
+                             @JsonProperty("salience") Double salience,
                              @JsonProperty("extendedProperties") Map<String, Object> extendedProperties) {
 
     }